+++ R/DataFunctions.R
@@ -1,5 +1,3 @@
-# library(sinew)
-# makeOxygen(getUniqueTrials)
 
 #' @title createTrial
 #' @description This function creates an object of class `scenario_list`
@@ -67,9 +65,6 @@ getScenario <- function (
 
 ) {
 
-  # n_cores      = parallel::detectCores() - 1L # old argument that is no longer needed
-  # "%dopar%" <- foreach::"%dopar%"
-
   response_rates           <- convertVector2Matrix(response_rates)
   colnames(response_rates) <- paste0("rr_", cohort_names)
 
@@ -92,8 +87,6 @@ getScenario <- function (
       response_rates = response_rates[, index_new],
       n_subjects     = n_subjects[index_new],
       n_trials       = n_trials)
-      # n_cores        = n_cores,
-      # seed           = seed)
 
   } else {
 
@@ -171,9 +164,6 @@ getScenario <- function (
                         previous_analyses = list(go_decisions   = previous_gos,
                                                  post_quantiles = NULL),
                         n_trials          = n_trials)
-                        # seed              = seed)
-
-  # class(scenario_data) <- "scenario_data"
 
   return (scenario_data)
 
@@ -397,15 +387,8 @@ simulateScenarios <- function (
 
   ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ###
 
-  # cat(format(Sys.time(), "%m/%d/%y"), " Simulating Scenarios\n", sep = "")
-
   scenario_list    <- vector(mode = "list", length = length(scenario_numbers))
   for (s in seq_along(scenario_numbers)) {
-
-    # start_time  <- Sys.time()
-    # out_message <- paste0(format(start_time, "   %H:%M"),
-    #                       " Simulating Scenario ", scenario_numbers[s], " ...")
-    # cat(out_message, rep(".", 33 + nchar(max(scenario_numbers)) - nchar(out_message)), sep = "")
 
     scenario_list[[s]] <- getScenario(
       n_subjects          = n_subjects_list[[s]],
@@ -413,10 +396,6 @@ simulateScenarios <- function (
       n_trials            = n_trials)
 
     scenario_list[[s]]$scenario_number <- scenario_numbers[s]
-
-    # cat(" Finished after ", round(Sys.time() - start_time, 1), " ",
-    #     units(Sys.time() - start_time), ".\n", sep = "")
-    # rm(start_time)
 
   }
 
@@ -432,21 +411,6 @@ is.scenario_list <- function (x) {
   if (missing(x)) stop ("Please provide an object for the argument 'x'")
 
   inherits(x, "scenario_list")
-
-  # ## x must be list
-  # if (!is.list(x)) {
-  #   warning ("x must be list")
-  #   return (FALSE)
-  # }
-  #
-  # is_scenario_data <- sapply(x, is.scenario_data)
-  #
-  # if (any(!is_scenario_data)) {
-  #   warning (paste0("The items of x with the following indices are not scenario_data", !which(is_scenario_data)))
-  #   return (FALSE)
-  # }
-  #
-  # return (TRUE)
 
 }
 
@@ -600,38 +564,6 @@ getRespondersNonParallel <- function (
 
 }
 
-
-getRespondersParallel <- function (
-
-  response_rates,
-  n_subjects,
-
-  n_trials,
-  n_cores = parallel::detectCores() - 1L
-  # seed    = as.numeric(Sys.time())
-
-) {
-
-  "%dopar%" <- foreach::"%dopar%"
-
-  ## somehow %dopar% does not default to %do% when no connection open
-  doParallel::registerDoParallel(n_cores)
-  on.exit(doParallel::stopImplicitCluster())
-  # if (n_cores > 1) {
-  #   doParallel::registerDoParallel(n_cores)
-  #   on.exit(doParallel::stopImplicitCluster())
-  # }
-  n_responders <- foreach::foreach(k = seq_len(n_trials),
-                                   .combine = rbind,
-                                   .export = c("getResponders")) %dopar% {
-                                     getResponders(response_rates = response_rates,
-                                                   n_subjects     = n_subjects)}
-                                                   # seed           = seed + k)}
-
-  return (n_responders)
-
-}
-
 #' @title continueRecruitment
 #' @md
 #' @description This function continues the recruitment of subjects for a set of scenarios
@@ -731,8 +663,6 @@ continueRecruitment <- function (
     stop (simpleError("The lengths of 'n_subjects_add_list' and 'decisions_list' must be equal"))
   }
 
-  # cat(format(Sys.time(), "%m/%d/%y"), " Continuing Recruitment\n", sep = "")
-
   scenario_list <- vector(mode = "list", length = length(decisions_list))
   names(scenario_list) <- paste0("scenario_", scenario_numbers)
   for (s in seq_along(scenario_list)) {
@@ -743,11 +673,6 @@ continueRecruitment <- function (
       stop (simpleError("Selected method_name not analyzed"))
 
     }
-
-    # start_time  <- Sys.time()
-    # out_message <- paste0(format(start_time, "   %H:%M"),
-    #                       " Simulating Scenario ", scenario_numbers[s], " ...")
-    # cat(out_message, rep(".", 33 + nchar(max(scenario_numbers)) - nchar(out_message)), sep = "")
 
     ## get new data, i.e. get new responders and new number of subjects per trial
 
@@ -777,14 +702,12 @@ continueRecruitment <- function (
     }
 
     n_trials <- decisions_list[[s]]$scenario_data$n_trials
-    # seed     <- decisions_list[[s]]$scenario_data$seed + n_trials + 1
 
     add_scenario <- getScenario(
       n_subjects     = n_subjects_add,
       response_rates = response_rates_new,
       cohort_names   = cohort_names_new,
       n_trials       = n_trials)
-      # seed           = seed)
 
     n_responders_add <- add_scenario$n_responders
     n_subjects_add   <- add_scenario$n_subjects
@@ -819,14 +742,9 @@ continueRecruitment <- function (
       previous_analyses = list(go_decisions   = previous_gos,
                                post_quantiles = decisions_list[[s]]$analysis_data$quantiles_list),
       n_trials          = n_trials)
-      # seed              = seed)
 
     scenario_list[[s]]$scenario_number <-
       decisions_list[[s]]$scenario_data$scenario_number
-
-    # cat(" Finished after ", round(Sys.time() - start_time, 1), " ",
-    #     units(Sys.time() - start_time), ".\n", sep = "")
-    # rm(start_time)
 
   }